chore: refactor get_peer_list into bindigs/runner/src/common.rs
