+++ bindings/runner/Cargo.toml
@@ -15,10 +15,14 @@ holochain_client_instrumented = { workspace = true }
 log = { workspace = true }
 anyhow = { workspace = true }
 url = { workspace = true }
+rand = {workspace = true}
 serde = { workspace = true }
 tokio = { workspace = true }
 holochain_types = { workspace = true }
 holochain_conductor_api = { workspace = true }
+kitsune2_api = { workspace = true }
+kitsune2_core = { workspace = true }
+holo_hash = { workspace = true, features = ["kitsune2"] }
 
 [lints]
 workspace = true
