+++ scenarios/remote_call_rate/Cargo.toml
@@ -15,7 +15,6 @@ holochain_wind_tunnel_runner = { workspace = true }
 remote_call_integrity = { workspace = true }
 kitsune2_api = { workspace = true }
 kitsune2_core = { workspace = true }
-holo_hash = { workspace = true, features = ["kitsune2"] }
 
 [build-dependencies]
 happ_builder = { workspace = true }
